Move getSynthetixContracts to utils. - This will be replaced once we're properly on kovan
